Creates more endpoints for comments: - breaks comment endpoints into their own controller w/ neccesary support classes - implements 'get all comments for x video' endpoint '/comments/get/{videoId}'
