+++ backend/src/main/java/com/smbirch/homemovies/mappers/CommentMapper.java
@@ -1,0 +1,8 @@
+package com.smbirch.homemovies.mappers;
+
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface CommentMapper {
+
+}
