+++ backend/src/main/java/com/smbirch/homemovies/services/VideoService.java
@@ -20,6 +20,4 @@ public interface VideoService {
     VideoResponseDto updateVideoTitle(VideoRequestDto videoRequestDto);
 
     VideoResponseDto updateVideoDescription(VideoRequestDto videoRequestDto);
-
-    CommentResponseDto postVideoComment(CommentRequestDto commentRequestDto);
 }
